+++ packages/mezon-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "mezon-sdk",
-  "version": "2.8.28",
+  "version": "2.8.29",
   "description": "Mezon typescript SDK.",
   "files": ["dist/**"],
   "main": "dist/cjs/index.js",
